+++ apps/client/public/CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog is only provided in English language.
 If you found some bugs or want us to add new feature, please do so via [our GitHub Tracker](https://github.com/sakilabs/openbts/issues/new) or send me an email [ririxi@sakilabs.com](mailto:ririxi@sakilabs.com)
 
+# 2026-05-18
+
+### 🚀 Enhancements
+
+- Photos uploaded in a NetWorks' station submission are now automatically added to the sibling station at the same location on approval
+- Added "Fetch from sibling" button in the Extra Identificators section. It pre-fills networks ID, networks name, and MNO name from the paired T-Mobile/Orange station at the same location
+
 # 2026-05-15
 
 ### 🚀 Enhancements